refactor(substrate-core): extract render and camera sink helpers (issue 428)
Desktop and test-bench register byte-for-byte identical render and
camera sinks. Move the shared scaffolding to a new
`aether-substrate-core::sinks` module exposing `build_render_sink(cap)`
and `build_camera_sink()`. Each helper returns the chassis-side state
(vertex buffer + triangle counter, or view-proj matrix) plus a ready-
to-register `SinkHandler`. Headless keeps its 4-line nop closures
inline since they are self-documenting.
